Encapsulate the supervisor api methods into a partial
Change-type: patch
